+++ rag/file_conversion_router/api.py
@@ -15,9 +15,3 @@ def convert_directory(input_dir: Union[str, Path], output_dir: Union[str, Path])
     2. Markdown (To clarify, this markdown includes additional tree structure of original markdown file)
     """
     process_folder(input_dir, output_dir)
-
-def main():
-    convert_directory(f"./rag/scraper/Scrape_ed/json_files", "./rag/scraper/Scrape_ed/output_files")
-
-if __name__ == "__main__":
-    main()
